Move Bootloader to pyanaconda.bootloader.base
Move the Bootloader class and the related code to a new file.
